+++ src/main/java/com/minhcraft/beyondbetatweaks/mixin/entity/PlayerMixin.java
@@ -1,0 +1,69 @@
+package com.minhcraft.beyondbetatweaks.mixin.entity;
+
+import com.minhcraft.beyondbetatweaks.config.ModConfig;
+import net.minecraft.world.entity.EntityType;
+import net.minecraft.world.entity.LivingEntity;
+import net.minecraft.world.entity.Pose;
+import net.minecraft.world.entity.player.Abilities;
+import net.minecraft.world.entity.player.Player;
+import net.minecraft.world.level.Level;
+import org.spongepowered.asm.mixin.Final;
+import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.Shadow;
+import org.spongepowered.asm.mixin.Unique;
+import org.spongepowered.asm.mixin.injection.At;
+import org.spongepowered.asm.mixin.injection.Inject;
+import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
+
+@Mixin(Player.class)
+public abstract class PlayerMixin extends LivingEntity {
+
+    /// Disable crawling code from https://github.com/AzaleaCatgirl99/Nostalgic-Tweaks-Renewed by @AzaleaCatgirl99
+
+    protected PlayerMixin(EntityType<? extends LivingEntity> entityType, Level level) {
+        super(entityType, level);
+    }
+
+    @Unique
+    protected boolean canPlayerFitWithinBlocksAndEntitiesWhen(Pose pose) {
+        return this.level().noCollision(this, this.getDimensions(pose).makeBoundingBox(this.position()).deflate(1.0E-7));
+    }
+
+    @Shadow @Final private Abilities abilities;
+
+    @Inject(method = "updatePlayerPose", at = @At("HEAD"), cancellable = true)
+    private void beyond_beta_tweaks$changeDefaultPoses(CallbackInfo ci) {
+        if (this.canPlayerFitWithinBlocksAndEntitiesWhen(Pose.SWIMMING)) {
+            Pose pose;
+            if (this.isFallFlying()) {
+                pose = Pose.FALL_FLYING;
+            } else if (this.isSleeping()) {
+                pose = Pose.SLEEPING;
+            } else if (this.isSwimming()) {
+                pose = Pose.SWIMMING;
+            } else if (this.isAutoSpinAttack()) {
+                pose = Pose.SPIN_ATTACK;
+            } else if (this.isShiftKeyDown() && !this.abilities.flying) {
+                pose = Pose.CROUCHING;
+            } else {
+                pose = Pose.STANDING;
+            }
+
+            Pose pose2;
+            if (!this.isSpectator() && !this.isPassenger() && !this.canPlayerFitWithinBlocksAndEntitiesWhen(pose)) {
+                if (this.canPlayerFitWithinBlocksAndEntitiesWhen(Pose.CROUCHING)) {
+                    pose2 = Pose.CROUCHING;
+                } else if (ModConfig.disableCrawling) {
+                    pose2 = Pose.STANDING;
+                } else {
+                    pose2 = Pose.SWIMMING;
+                }
+            } else {
+                pose2 = pose;
+            }
+
+            this.setPose(pose2);
+        }
+        ci.cancel();
+    }
+}